improve generator

+++ src/framework.ts
@@ -50,50 +50,57 @@ export const runCreateNuxt = (callback:Function) =>{
 export const prepareNest = (callback:Function)=>{
     const targetfolder =config.backendFolder
     log.info(`creating backend project ${targetfolder}`)
-    exec(`cd ${targetfolder};pnpm install --save @nestjs/swagger @nestjs/mongoose mongoose  ajv ajv-formats @nestjs/config`,async (error, stdout, stderr)=>{
-        // log.info(`dependency installed`)
-        if(!error){
-            exec(`pnpm install ajv ajv-formats axios json-schema`, (error, stdout, stderr)=>{                
-            const eta = new Eta({views: constants.templatedir});              
-            const variables={
-                backendPort:config.backendPort,
-                mongoConnectStr:config.mongoConnectStr
-            }            
-            const txtEnv = eta.render('./nest/nest.env.eta', variables);                
-            const txtMain = eta.render('./nest/nest.main.eta', variables);   
+    if(!fs.existsSync(`${targetfolder}/.env`)){
 
-            fs.writeFileSync(`${targetfolder}/.env`, txtEnv);
-            fs.writeFileSync(`${targetfolder}/src/main.ts`, txtMain);            
-            const tsconfigpath = process.cwd()+'/'+`${targetfolder}/tsconfig.json`
-            const tsconfig = require(tsconfigpath)
-            tsconfig.compilerOptions.esModuleInterop=true
-            tsconfig.compilerOptions.resolveJsonModule=true
-            fs.writeFileSync(tsconfigpath, JSON.stringify(tsconfig));
+    
+        exec(`cd ${targetfolder};pnpm install --save axios @darkwolf/base64url json-schema @wearenova/mongoose-tenant @nestjs/swagger @nestjs/mongoose mongoose  ajv ajv-formats @nestjs/config`,async (error, stdout, stderr)=>{
+            // log.info(`dependency installed`)
+            if(!error){
+                            
+                const eta = new Eta({views: constants.templatedir});              
+                const variables={
+                    backendPort:config.backendPort,
+                    mongoConnectStr:config.mongoConnectStr
+                }            
+                const txtEnv = eta.render('./nest/nest.env.eta', variables);                
+                const txtMain = eta.render('./nest/nest.main.eta', variables);   
 
-            log.info("nest project completed")                
-            callback()
+                fs.writeFileSync(`${targetfolder}/.env`, txtEnv);
+                fs.writeFileSync(`${targetfolder}/src/main.ts`, txtMain);            
+                const tsconfigpath = process.cwd()+'/'+`${targetfolder}/tsconfig.json`
+                const tsconfig = require(tsconfigpath)
+                tsconfig.compilerOptions.esModuleInterop=true
+                tsconfig.compilerOptions.resolveJsonModule=true
+                fs.writeFileSync(tsconfigpath, JSON.stringify(tsconfig));
+
+                log.info("nest project completed")                
+                callback()
+            
+            } else{
+            log.error(stderr)
+            throw error
+            }
         })
-        } else{
-        log.error(stderr)
-        throw error
-        }
-    })
+    }else{
+        log.info(`${targetfolder}/.env exists, skip regenerate environment`)
+        callback()
+    }
 }
 //prepare nuxt project for simpleapp generator
 export const prepareNuxt = (callback:Function)=>{
     const targetfolder = config.frontendFolder
     if(!fs.existsSync(`${targetfolder}/.env`)){
         //asume no environment. prepare now
-        exec(`cd ${targetfolder};pnpm install;pnpm install -D @types/node @vueuse/nuxt @sidebase/nuxt-auth @vueuse/core nuxt-security prettier @nuxtjs/tailwindcss`, (error, stdout, stderr)=>{                
+        exec(`cd ${targetfolder};pnpm install;pnpm install -D @nuxt/ui @types/node @vueuse/nuxt @sidebase/nuxt-auth @vueuse/core nuxt-security prettier `, (error, stdout, stderr)=>{                
             //;pnpm install    
             console.log(error, stdout, stderr)
-                exec(`cd ${targetfolder};pnpm install --save ajv dotenv @fullcalendar/core @fullcalendar/vue3 quill uuid ajv-formats primeflex primeicons prettier primevue axios json-schema mitt @simitgroup/simpleapp-vue-component@latest`, (error, stdout, stderr)=>{                
+                exec(`cd ${targetfolder};pnpm install --save  @darkwolf/base64url @nuxt/ui ajv dotenv @fullcalendar/core @fullcalendar/vue3 quill uuid ajv-formats primeflex primeicons prettier primevue axios json-schema mitt @simitgroup/simpleapp-vue-component@latest`, (error, stdout, stderr)=>{                
                 console.log(error, stdout, stderr)
                 
                 fs.mkdirSync(`${targetfolder}/assets/css/`,{recursive:true})
                 fs.mkdirSync(`${targetfolder}/layouts`,{recursive:true})
                 fs.mkdirSync(`${targetfolder}/components`,{recursive:true})
-                fs.mkdirSync(`${targetfolder}/server/api`,{recursive:true})
+                fs.mkdirSync(`${targetfolder}/server/api/[xorg]`,{recursive:true})
                 fs.mkdirSync(`${targetfolder}/pages`,{recursive:true})
                 fs.mkdirSync(`${targetfolder}/plugins`,{recursive:true})
                 const eta = new Eta({views: `${constants.templatedir}/nuxt`});              
@@ -108,7 +115,7 @@ export const prepareNuxt = (callback:Function)=>{
                     './components.crudsimple.vue.eta':'components/CrudSimple.vue',            
                     './components.debugdocdata.vue.eta':'components/DebugDocumentData.vue',            
                     './layouts.default.vue.eta':'layouts/default.vue',
-                    './server.api.ts.eta':'server/api/[...].ts',
+                    './server.api.ts.eta':'server/api/[xorg]/[...].ts',
                     './nuxt.config.ts.eta':'nuxt.config.ts',
                     './pages.index.vue.eta':'pages/index.vue',
                     './plugins.simpleapp.ts.eta':'plugins/simpleapp.ts',
@@ -143,7 +150,7 @@ export const prettyNuxt = ()=>{
                         
 }
 export const prettyNest = ()=>{
-    exec(`cd ${config.backendFolder};npx run format `)
+    exec(`cd ${config.backendFolder};npm run format`)
 }
 
 export const prepareOpenApiClient = () => {
